+++ Editor/IconPostProcessor.cs
@@ -14,15 +14,18 @@ namespace IndustryCSE.Tool.ProductConfigurator.Editor
             foreach (var asset in importedAssets)
             {
                 if (!asset.EndsWith(".png", StringComparison.OrdinalIgnoreCase)) continue;
-                if (Directory.GetParent(asset).ToString() !=
-                    EditorCore.VariantSetIconPath) continue;
-                var texture = AssetDatabase.LoadAssetAtPath<Texture2D>(asset);
+
+                var fullIconPath = Path.Combine(Directory.GetParent(Application.dataPath).FullName,
+                    EditorCore.VariantSetIconPath);
+                
+                if (!string.Equals(Directory.GetParent(asset).FullName, fullIconPath)) continue;
                 string id = Path.GetFileNameWithoutExtension(asset);
                 foreach (var configurationAsset in AssetDatabase.FindAssets($"t:{nameof(VariantAsset)}"))
                 {
                     var configurationOption = AssetDatabase.LoadAssetAtPath<VariantAsset>(
                         AssetDatabase.GUIDToAssetPath(configurationAsset));
                     if (!string.Equals(configurationOption.UniqueIdString, id)) continue;
+                    var texture = AssetDatabase.LoadAssetAtPath<Texture2D>(asset);
                     configurationOption.icon = texture;
                     EditorUtility.SetDirty(configurationOption);
                 }