+++ Editor/Variant Set Editor/CombinationVariantSetEditor.cs
@@ -1,5 +1,6 @@
 using IndustryCSE.Tool.ProductConfigurator.ScriptableObjects;
 using UnityEditor;
+using IndustryCSE.Tool.ProductConfigurator.Runtime;
 
 namespace IndustryCSE.Tool.ProductConfigurator.Editor
 {