+++ Editor/Variant Set Editor/TransformVariantSetEditor.cs
@@ -1,7 +1,5 @@
-using IndustryCSE.Tool.ProductConfigurator.ScriptableObjects;
 using UnityEditor;
-using UnityEngine;
-using System;
+using IndustryCSE.Tool.ProductConfigurator.Runtime;
 
 namespace IndustryCSE.Tool.ProductConfigurator.Editor
 {