+++ Runtime/Base Class/VariantBase.cs
@@ -1,8 +1,9 @@
 using System;
 using System.Collections.Generic;
 using IndustryCSE.Tool.ProductConfigurator.ScriptableObjects;
+using UnityEngine;
 
-namespace IndustryCSE.Tool.ProductConfigurator
+namespace IndustryCSE.Tool.ProductConfigurator.Runtime
 {
     [Serializable]
     public struct ConditionalVariantData
@@ -17,6 +18,11 @@ namespace IndustryCSE.Tool.ProductConfigurator
         public VariantAsset variantAsset;
 
         public List<ConditionalVariantData> conditionalVariants;
+        
+        #if UNITY_EDITOR
+        [HideInInspector]
+        public bool FoldoutInspector = true;
+        #endif
     }
 }